Do not try to remove daemon files when the tree is still not initialized

+++ index.js
@@ -33,17 +33,18 @@ class Etc {
     }
 
     const xConfig = this.load ('xcraft');
-
-    /* Clean obsolete daemon files */
-    const isRunning = require ('is-running');
-    const runDir = path.join (xConfig.xcraftRoot, `var/run`);
-    xFs
-      .ls (runDir, /^xcraftd.[0-9]+$/)
-      .map (name => path.join (runDir, name))
-      .filter (
-        file => !isRunning (parseInt (file.replace (/.*(\.[0-9]+$)/, '$1')))
-      )
-      .forEach (fs.unlinkSync);
+    if (xConfig) {
+      /* Clean obsolete daemon files */
+      const isRunning = require ('is-running');
+      const runDir = path.join (xConfig.xcraftRoot, `var/run`);
+      xFs
+        .ls (runDir, /^xcraftd.[0-9]+$/)
+        .map (name => path.join (runDir, name))
+        .filter (
+          file => !isRunning (parseInt (file.replace (/.*(\.[0-9]+$)/, '$1')))
+        )
+        .forEach (fs.unlinkSync);
+    }
   }
 
   /**
